[migration-tools] Don't assume that every release repository has a tracks.yaml. (#32804)
A malformed release repository could be missing its tracks.yaml
configuration file. When this happens it should raise a ValueError as
with any other failure during the release process so that the migration
can continue.

---

Some guards are put in place because of things you know could happen.
Others are added because reality surpassed even your wildest
expectations.