[migration-tools] Catch GitHub errors during migration (#36566)
This is desirable because a migration may progress substantially before
encountering a repository that needs to be created, and the GitHub key
may not have that specific permission. If we don't catch the exception
here, the entire migration will need to be started again because the
progress will be lost.