Update patch config when copying refs.
Bloom creates `patches/` branches for each step in the release
generation process. These branches store patches and meta-information.

In most cases, bloom will do the "right thing" when needing to update
patch configurations after a ref is copied. However, in the case where a
new rosdistro is being added for a branch that has patches, the patch
configuration is reset when automatically updating the patch
configuration. So we need to reach into the patch configuration and
update the parent field so it points to the copied ref rather than the
original.